+++ src/Zope2/Startup/handlers.py
@@ -119,11 +119,6 @@ def root_handler(config):
             L.append(d)
     Products.__path__[:] = L
 
-    # Augment the set of MIME types:
-    if config.mime_types:
-        from zope.contenttype import add_files
-        add_files(config.mime_types)
-
     # if no servers are defined, create default http server and ftp server
     if not config.servers:
         config.servers = []