+++ doc/CHANGES.rst
@@ -44,6 +44,9 @@ Features Added
 Restructuring
 +++++++++++++
 
+- Removed `mime-types` option from `zope.conf`. You can use the `add_files`
+  API from `zope.contenttype` instead.
+
 - Removed `OFS.ZDOM`.
 
 - Removed the last remaining code to support `SOFTWARE_HOME` and `ZOPE_HOME`.